Moved template functions to their header files
Silly me I was getting a linker error due to the template functions not being defined.
